+++ src/ft/ft_puthex.c
@@ -6,13 +6,13 @@
 /*   By: fjacquem <fjacquem@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2014/07/05 09:53:57 by fjacquem          #+#    #+#             */
-/*   Updated: 2018/03/12 16:40:40 by fxst1            ###   ########.fr       */
+/*   Updated: 2018/04/14 15:41:15 by fjacquem         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include <libft.h>
 
-static void		ft_puthex_rec(unsigned int number, const int fd)
+static void		ft_puthex_rec(long number, const int fd)
 {
 	char	c;
 	char	*map;
@@ -27,7 +27,7 @@ static void		ft_puthex_rec(unsigned int number, const int fd)
 	}
 }
 
-void			ft_puthex_fd(unsigned int number, const int fd)
+void			ft_puthex_fd(long number, const int fd)
 {
 	ft_putstr_fd("0x", fd);
 	if (number)
